Use SingleVariantOperatorBase as the class to perform operations on every new aligned Variant object produced by the scan_and_operate() function. Every SingleVariantOperatorBase object is supposed to define its version of the operate() function which will be called by the scan_and_operate() function whenever a new aligned Variant object is produced. The arguments to operate() are the new Variant object produced and the VariantQueryConfig object. The Operator objects are free to initialize any members as required before the scan_and_operate function is called. The Variant object passed as argument to the operate() function of a VariantOperator can go out of scope as soon as the operate() function completes. Hence, VariantOperator objects CANNOT just store pointers to data associated with this Variant objects and expect the pointers to be valid when operate is called again. The Operator() MUST copy the Variant object for correctness if the information from the Variant will be used in the future. Other edits: 1. Sample driver code for VariantOperators, including one for GA4GH (merged REF, ALT and re-ordered ALT) 2. LUTs now deal with int64_t exclusively 3. Variant objects can now store fields that are common across Calls - e.g. merged REF,ALT etc. Could be useful as new Operators are added. 4. Remap wrapper makes it easy for remap_ functions to re-order PL, AD fields, irrespective of the target data structure.

+++ variant/src/lut.cc
@@ -11,7 +11,7 @@ LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_order
 }
 
 template<bool inputs_2_merged_LUT_is_input_ordered, bool merged_2_inputs_LUT_is_input_ordered>
-LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_ordered>::LUTBase(unsigned numInputGVCFs, unsigned numMergedFields)
+LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_ordered>::LUTBase(int64_t numInputGVCFs, int64_t numMergedFields)
 {
   m_num_input_vcfs = numInputGVCFs;
   m_num_merged_fields = numMergedFields;
@@ -32,14 +32,14 @@ void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_
 }
 
 template<bool inputs_2_merged_LUT_is_input_ordered, bool merged_2_inputs_LUT_is_input_ordered>
-void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_ordered>::reset_vector(vector<int>& vec, unsigned from)
+void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_ordered>::reset_vector(vector<int64_t>& vec, int64_t from)
 {
   for(auto i=from;i<vec.size();++i)
     vec[i] = lut_missing_value;
 }
 
 template<bool inputs_2_merged_LUT_is_input_ordered, bool merged_2_inputs_LUT_is_input_ordered>
-void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_ordered>::resize_and_reset_vector(vector<int>& vec, unsigned new_size)
+void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_ordered>::resize_and_reset_vector(vector<int64_t>& vec, int64_t new_size)
 {
   auto old_size = vec.size();
   if(new_size > old_size)
@@ -51,7 +51,7 @@ void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_
 
 template<bool inputs_2_merged_LUT_is_input_ordered, bool merged_2_inputs_LUT_is_input_ordered>
 void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_ordered>::resize_and_reset_lut
-(vector<vector<int>>& lut, unsigned new_lut_size, unsigned new_vector_size, unsigned& numRowsVar, unsigned& numColsVar)
+(vector<vector<int64_t>>& lut, int64_t new_lut_size, int64_t new_vector_size, int64_t& numRowsVar, int64_t& numColsVar)
 {
   auto old_lut_size = lut.size();
   if(new_lut_size > old_lut_size)
